CodeQL fixes
